iter1.2 | Добавлены методы хэширования и проверки паролей, авторизация через токен, обновлены зависимости
Добавлены методы для хэширования и проверки паролей пользователя в модели User, обработчик авторизации по токену, новые методы репозиториев для получения пользователей по логину и ID. Обновлены зависимости в go.mod и go.sum, добавлены вспомогательные функции для работы с ключами JWT.

Changed files:
- internal/models/user.go
- internal/payloads/register.go
- internal/repositories/user.go
- internal/configuration/parse.go
- internal/token/middleware.go
- internal/token/configuration.go
- go.mod
- go.sum
